Fixed reading database
missing
- add to database
- remove from database
-edit database
