+++ src/app/projects/page.tsx
@@ -1,18 +1,26 @@
 'use client';
 
 import { useEffect, useState } from "react";
-import { getAllProjects } from '@/lib/firebase/firebaseOperations';
+import { getAllProjects, Project } from '@/lib/firebase/firebaseOperations';
 
 export default function Projects() {
-
-  let x;
+  const [projects, setProjects] = useState<Project[]>([]);
   
+  
+ // The empty array ensures useEffect runs only once on mount
+
   useEffect(() => {
-    x = getAllProjects();
-  }, []);  // The empty array ensures useEffect runs only once on mount
+    const fetchProjects = async () => {
+      const data = await getAllProjects();
+      setProjects(data); // Set the state with the fetched projects
+    };
+
+    fetchProjects(); // Call the function to fetch projects
+  }, []);
+
 
   return (
     <>Projects page
-    <pre>{JSON.stringify(x, null, 2)}</pre></>
+    <pre>{JSON.stringify(projects, null, 4)}</pre></>
   );
 }
